Do the hack only if .gitattributes exists

+++ checkout.sh
@@ -5,7 +5,9 @@ set -eo pipefail
 git fetch
 git switch -d origin/HEAD
 
-cp ../gitattributes-hack .gitattributes
-git status
-git restore .gitattributes
+if [ -f .gitattributes ]; then
+	cp ../gitattributes-hack .gitattributes
+	git status
+	git restore .gitattributes
+fi
 git status
